Read events paging abstraction
Changed read events to hide the paging from the consumer. Instead of
returning PagedEventsResult it returns an Enumerable object that makes
the requests lazily. This should make the gem easier to use for the majority of
developers.

Because the page URLs are no longer exposed to the user, the get_events_page
method was removed from the client.

Within this commit all requests were altered to report the version of the gem as
the User-Agent.