Made read events partially lazily evaluated
The first page of events is requested eagerly so that common errors such
as expired tokens and argument errors will happen in-line. Requests for
subsequent pages happen lazily.

This is an attempt to strike a balance between ease-of-use of making all
the requests in-line and the efficiency of loading each page as needed.