Expose validation errors more easily
Previously the caller would have to parse the JSON response themselves.
As it is part of the API's contract we should do this work on their
behalf.

However, we should not propagate any parsing errors upwards as this will
be bad for callers. Therefore, we will return an empty set of errors if
an unexpected response, or response we cannot parse, is found.