Less restrictive hashie dependency
Many other gems do not allow a dependency on hashie >= 3, therefore to
avoid conflicting dependencies we have loosened our requirements for the
hashie gem to include >= 2.1

This required some changes as we relied on features from hashie >= 3 in
a couple of places.

The CI script was also augmented to run tests against hashie ~> 2.1.0 as
well as the most recent ~> 3.0 release.

Supercedes #5